Decouple ctrl-drag code to its own file
